Add doc input `Number::class`.
