+++ tests/Input/Number/AttributeTest.php
@@ -153,26 +153,6 @@ final class AttributeTest extends TestCase
         );
     }
 
-    public function testMax(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="number-6582f2d099e8b" type="number" max="value">
-            HTML,
-            Number::widget()->id('number-6582f2d099e8b')->max('value')->render()
-        );
-    }
-
-    public function testMin(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="number-6582f2d099e8b" type="number" min="value">
-            HTML,
-            Number::widget()->id('number-6582f2d099e8b')->min('value')->render()
-        );
-    }
-
     public function testName(): void
     {
         Assert::equalsWithoutLE(
@@ -202,37 +182,7 @@ final class AttributeTest extends TestCase
             Number::widget()->id('number-6582f2d099e8b')->readonly()->render()
         );
     }
-
-    public function testRender(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="number-6582f2d099e8b" type="number">
-            HTML,
-            Number::widget()->id('number-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testRequired(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="number-6582f2d099e8b" type="number" required>
-            HTML,
-            Number::widget()->id('number-6582f2d099e8b')->required()->render()
-        );
-    }
-
-    public function testStep(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="number-6582f2d099e8b" type="number" step="1">
-            HTML,
-            Number::widget()->id('number-6582f2d099e8b')->step(1)->render()
-        );
-    }
-
+    
     public function testStyle(): void
     {
         Assert::equalsWithoutLE(
